requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.8
pandas==2.2.2
openpyxl==3.1.5
gunicorn==22.0.0
uvicorn==0.30.6
